Refactor firestore integration. Adds angular service GameRoomService.
